Moved Exception class and Enums in the corresponding packages
