+++ src/main/java/com/github/jreddit/exception/InvalidCookieException.java
@@ -1,4 +1,4 @@
-package com.github.jreddit;
+package com.github.jreddit.exception;
 
 import java.io.IOException;
